pipeline_orchestrator/backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Task queue
celery==5.3.6
redis==5.0.1
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Google Cloud
google-cloud-bigquery==3.14.1
google-auth==2.26.2

# Data processing
pandas==2.2.0
numpy==1.26.3

# Scheduling
croniter==2.0.1
APScheduler==3.10.4

# Monitoring & Logging
python-json-logger==2.0.7

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
httpx==0.26.0


